exposure-platform-rebuild 2/backend/requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.1
SQLAlchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.2
PyJWT==2.8.0
boto3==1.34.9
numpy==1.26.4
scipy==1.13.1
pandas==2.2.2
geojson==3.1.0
python-multipart==0.0.9
